新增了服务器启动脚本，目前浏览器前端可以正确给后端发送指令

+++ backend/src/main/resources/db/migration/V4__create_commands.sql
@@ -1,7 +1,7 @@
 -- 创建 commands 表，字段与 Command 实体严格一致
 CREATE TABLE commands (
     id BIGSERIAL PRIMARY KEY,
-    terminal_id BIGINT NOT NULL,
+    terminal_id INTEGER NOT NULL,
     cmd_num INTEGER NOT NULL, -- 对应 CommandType 枚举的编号
     created_at TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
     CONSTRAINT fk_command_terminal
